refactor(app): extract question draft cache and option-focus helpers
Move the per-request draft cache (with the QuestionAnswer/DraftAnswer answer
types its value shape is defined in terms of) into question-draft.ts, and the
keepVisibleInQuestionOptions/focusWithoutScrollingTimeline DOM helpers into
question-option-focus.ts. Pure extraction, no behavior change; the dock imports
both back. Dependency flows one way (dock to the new modules).
